created users model, cleaned up and consolidated migrations, created new tutor/tutee models and marked old ones as such for temporary reference, updated activerecord associations to be more correct (took several hours of reading docs and examples but I think it makes way more sense now?), updated seeds.rb to match new schema, added back in all the deleted devise views to default to, or at least for reference
