+++ app/controllers/requests_controller.rb
@@ -16,7 +16,7 @@ class RequestsController < ApplicationController
     @tutee = Tutee.find_by_id(params[:tutee_id])
     @course_array = Admin.course_list
     @meeting_time = [["1 hour",1], ["1.5 hours",1.5], ["2 hours",2]]
-    @has_priority = Admin.priority_list_contains? @tutee.sid
+    @has_priority = Admin.priority_list_contains? @tutee.sid #TODO: replace SID instances with email fields
     @tutee_most_recent_request = @tutee.requests.order('created_at ASC').last
     if @tutee_most_recent_request
       status = @tutee_most_recent_request.status
