+++ app/models/evaluation.rb
@@ -1,9 +1,6 @@
 class Evaluation < ApplicationRecord
   include Friendlyable
   has_one :meeting
-  has_one :request, through: :meeting
-  has_one :tutor, through: :meeting
-  has_one :tutee, through: :request
   has_many :question
 
   validates :status, presence: true, inclusion: { in: %w(Pending Complete), message: "Must be valid status"}, on: :update, :if => :took_place
