+++ app/models/question.rb
@@ -1,6 +1,2 @@
 class Question < ApplicationRecord
-  belongs_to :evaluation
-  belongs_to :question_template
-
-
 end
