+++ config/routes.rb
@@ -1,5 +1,6 @@
 Rails.application.routes.draw do
 
+  devise_for :users
   devise_for :tutors, controllers: {registrations: 'tutors/registrations'}, skip:['sessions']
   devise_for :tutees, skip:['registrations','sessions']
   devise_scope :tutor do
