--- db/migrate/20210419022628_add_is_done.rb
@@ -1,5 +1,0 @@
- class AddIsDone < ActiveRecord::Migration[5.2]
-  def change
-    add_column :meetings, :is_done, :boolean, default: false
-  end
-end
