--- db/migrate/20210422225724_add_signups_allowed_to_admins.rb
@@ -1,5 +1,0 @@
-class AddSignupsAllowedToAdmins < ActiveRecord::Migration[5.2]
-  def change
-  	add_column :admins, :signups_allowed, :boolean, :default => true
-  end
-end
