--- db/migrate/20210429013521_add_tutor_types_to_admins.rb
@@ -1,5 +1,0 @@
-class AddTutorTypesToAdmins < ActiveRecord::Migration[5.2]
-  def change
-    add_column :admins, :tutor_types, :text, :default=> "this is default text"
-  end
-end
