--- db/migrate/20210430031022_add_priority_list_to_admins.rb
@@ -1,5 +1,0 @@
-class AddPriorityListToAdmins < ActiveRecord::Migration[5.2]
-  def change
-    add_column :admins, :priority_list, :integer, :array=>true
-  end
-end
