--- db/migrate/20210430065313_add_status_to_request.rb
@@ -1,5 +1,0 @@
-class AddStatusToRequest < ActiveRecord::Migration[5.2]
-  def change
-    add_column :requests, :status, :string, default: "open"
-  end
-end
