--- db/migrate/20210503003432_add_course_list_to_admins.rb
@@ -1,5 +1,0 @@
-class AddCourseListToAdmins < ActiveRecord::Migration[5.2]
-  def change
-    add_column :admins, :course_list, :string, :array=>true
-  end
-end
