+++ db/migrate/20210903210500_create_admins.rb
@@ -1,0 +1,12 @@
+class CreateAdmins < ActiveRecord::Migration[5.2]
+  def change
+    create_table :admins do |t|
+      t.string :password_digest
+      t.string :course_list, array: true, default: ["CS10", "CS61A", "CS61B", "CS61C", "CS70", "CS88", "EE16A", "EE16B", "DATA8", "UPPERDIV"]
+      t.text :tutor_types, default: "CSM (8-10 hours)\r\nTA (12 hours)\r\nAcademic Intern (36 hours)\r\nTutor (12 hours)"
+      t.integer :priority_list, array: true, default: [333333] #TODO: change this to use emails instead of SIDs
+      t.boolean :signups_allowed, default: true
+      t.timestamps
+    end
+  end
+end
